requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiolimiter==1.2.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   claude-agent-sdk
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.4.0
    # via
    #   cyclopts
    #   jsonlines
    #   jsonschema
    #   referencing
authlib==1.6.5
    # via fastmcp
beautifulsoup4==4.14.2
    # via paper-search-mcp
build==1.3.0
    # via pip-tools
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
claude-agent-sdk==0.1.1
    # via -r requirements.in
click==8.3.0
    # via
    #   pip-tools
    #   typer
    #   uvicorn
    #   waybackpy
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
coverage[toml]==7.10.7
    # via pytest-cov
cryptography==46.0.2
    # via authlib
cyclopts==3.24.0
    # via fastmcp
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22.2
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.0
    # via fastmcp
fastmcp==2.12.4
    # via paper-search-mcp
feedparser==6.0.12
    # via paper-search-mcp
greenlet==3.2.4
    # via playwright
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   fastmcp
    #   mcp
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
isodate==0.7.2
    # via openapi-core
jsonlines==4.0.0
    # via -r requirements.in
jsonschema==4.25.1
    # via
    #   -r requirements.in
    #   mcp
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2025.9.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.12.0
    # via openapi-spec-validator
lxml==6.0.2
    # via paper-search-mcp
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via werkzeug
mcp[cli]==1.16.0
    # via
    #   -r requirements.in
    #   claude-agent-sdk
    #   fastmcp
    #   paper-search-mcp
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via openapi-core
mypy==1.18.2
    # via -r requirements.in
mypy-extensions==1.1.0
    # via mypy
numpy==2.3.3
    # via pandas
openapi-core==0.19.5
    # via fastmcp
openapi-pydantic==0.5.1
    # via fastmcp
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.2
    # via openapi-core
packaging==25.0
    # via
    #   build
    #   pytest
pandas==2.3.3
    # via -r requirements.in
paper-search-mcp==0.1.3
    # via -r requirements.in
parse==1.20.2
    # via openapi-core
pathable==0.4.4
    # via jsonschema-path
pathspec==0.12.1
    # via mypy
pip-tools==7.5.1
    # via -r requirements.in
playwright==1.55.0
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pycparser==2.23
    # via cffi
pydantic[email]==2.11.10
    # via
    #   -r requirements.in
    #   fastmcp
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.11.0
    # via mcp
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via
    #   pytest
    #   rich
pypdf2==3.0.1
    # via paper-search-mcp
pyperclip==1.11.0
    # via fastmcp
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
pytest-asyncio==1.2.0
    # via -r requirements.in
pytest-cov==7.0.0
    # via -r requirements.in
pytest-mock==3.15.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   fastmcp
    #   mcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pytz==2025.2
    # via pandas
pywin32==311
    # via mcp
pyyaml==6.0.3
    # via jsonschema-path
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
    #   types-jsonschema
requests==2.32.5
    # via
    #   jsonschema-path
    #   paper-search-mcp
    #   waybackpy
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==14.1.0
    # via
    #   -r requirements.in
    #   cyclopts
    #   fastmcp
    #   rich-rst
    #   typer
rich-rst==1.3.1
    # via cyclopts
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruff==0.13.3
    # via -r requirements.in
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
sse-starlette==3.0.2
    # via mcp
starlette==0.48.0
    # via mcp
structlog==25.4.0
    # via -r requirements.in
tenacity==9.1.2
    # via -r requirements.in
typer==0.19.2
    # via mcp
types-jsonschema==4.25.1.20250822
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   exceptiongroup
    #   mypy
    #   openapi-core
    #   pydantic
    #   pydantic-core
    #   pyee
    #   pytest-asyncio
    #   referencing
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   requests
    #   waybackpy
uvicorn==0.37.0
    # via mcp
waybackpy==3.0.6
    # via -r requirements.in
werkzeug==3.1.1
    # via openapi-core
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
